reorganizing the folders
this might cause some merge conflicts bc I moved a lot of files around so just lmk
