+++ App.js
@@ -10,10 +10,10 @@ import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import MainTabScreen from './screens/MainTabScreen';
-import LoginScreen from './screens/LoginScreen';
-import RegistrationScreen from './screens/RegistrationScreen';
-import WelcomeScreen from './screens/WelcomeScreen';
-import InfoScreen from './screens/InfoScreen';
+import LoginScreen from './screens/WelcomePages/LoginScreen';
+import RegistrationScreen from './screens/WelcomePages/RegistrationScreen';
+import WelcomeScreen from './screens/WelcomePages/WelcomeScreen';
+import InfoScreen from './screens/WelcomePages/InfoScreen';
 
 const Stack = createNativeStackNavigator();
