+++ screens/AnimalTiles/AnimalTile2.js
