+++ screens/MainTabScreen.js
@@ -5,14 +5,14 @@ import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {StyleSheet, Image} from 'react-native';
 import HomeScreen from './HomeScreen';
 import LikedScreen from './LikedScreen';
-import ShelterScreen from './ShelterScreen';
+import ShelterScreen from './SearchPage/ShelterScreen';
 import ProfileScreen from './ProfileScreen';
-import AnimalTile from './AnimalTile';
-import AnimalTile2 from './AnimalTile2';
-import pawImage from './2.png';
-import likeImage from './12.png';
-import userImage from './3.png';
-import searchImage from './4.png';
+import AnimalTile from './AnimalTiles/AnimalTile';
+import AnimalTile2 from './AnimalTiles/AnimalTile2';
+import pawImage from '../icons/2.png';
+import likeImage from '../icons/12.png';
+import userImage from '../icons/3.png';
+import searchImage from '../icons/4.png';
 
 const Tab = createMaterialBottomTabNavigator();
 const LikedStack = createNativeStackNavigator();
